GCS_MAVLINK_Rover: change return of vfr_hud_throttle() from int16 to float

+++ Rover/GCS_Mavlink.cpp
@@ -142,7 +142,7 @@ void GCS_MAVLINK_Rover::send_servo_out()
         receiver_rssi());
 }
 
-int16_t GCS_MAVLINK_Rover::vfr_hud_throttle() const
+float GCS_MAVLINK_Rover::vfr_hud_throttle() const
 {
     return rover.g2.motors.get_throttle();
 }
