+++ Rover/GCS_Mavlink.h
@@ -51,7 +51,7 @@ private:
     MAV_MODE base_mode() const override;
     MAV_STATE vehicle_system_status() const override;
 
-    int16_t vfr_hud_throttle() const override;
+    float vfr_hud_throttle() const override;
 
     void send_rangefinder() const override;
